# backends/booking-service/swagger.json
{"swagger": "2.0", "basePath": "/", "paths": {"/booking/book": {"post": {"responses": {"200": {"description": "Success", "schema": {"$ref": "#/definitions/BookingResponse"}}}, "operationId": "post_booking", "parameters": [{"name": "payload", "required": true, "in": "body", "schema": {"$ref": "#/definitions/BookingInput"}}, {"name": "X-Fields", "in": "header", "type": "string", "format": "mask", "description": "An optional fields mask"}], "tags": ["booking"]}}, "/booking/rooms": {"get": {"responses": {"200": {"description": "Success", "schema": {"type": "array", "items": {"$ref": "#/definitions/Room"}}}}, "operationId": "get_room_list", "parameters": [{"name": "X-Fields", "in": "header", "type": "string", "format": "mask", "description": "An optional fields mask"}], "tags": ["booking"]}}, "/booking/rooms/{room_id}": {"parameters": [{"name": "room_id", "in": "path", "required": true, "type": "string"}], "get": {"responses": {"200": {"description": "Success", "schema": {"$ref": "#/definitions/Room"}}}, "operationId": "get_room", "parameters": [{"name": "X-Fields", "in": "header", "type": "string", "format": "mask", "description": "An optional fields mask"}], "tags": ["booking"]}}}, "info": {"title": "API", "version": "1.0"}, "produces": ["application/json"], "consumes": ["application/json"], "tags": [{"name": "booking", "description": "Booking operations"}], "definitions": {"Room": {"required": ["available", "id", "number"], "properties": {"id": {"type": "string"}, "number": {"type": "integer"}, "available": {"type": "boolean"}}, "type": "object"}, "BookingInput": {"required": ["check_in", "check_out", "email", "name", "phone", "room_number"], "properties": {"check_in": {"type": "string", "description": "Check-in date (ISO 8601)"}, "check_out": {"type": "string", "description": "Check-out date (ISO 8601)"}, "email": {"type": "string"}, "name": {"type": "string"}, "phone": {"type": "string"}, "room_number": {"type": "integer"}}, "type": "object"}, "BookingResponse": {"required": ["booking_id", "status"], "properties": {"status": {"type": "string"}, "booking_id": {"type": "string"}}, "type": "object"}}, "responses": {"ParseError": {"description": "When a mask can't be parsed"}, "MaskError": {"description": "When any error occurs on mask"}}}